Create waterBill
This code calculate water bill based in this problem: 

"The local water supply company has an incentive program to save water, and has changed the prices of its supply so that, proportionally, those customers who consume less water pay less per cubic meter. Every customer pays a monthly subscription of R$ 9.00, which includes an excess of 10 m3 of water. That is, for any consumption between 0 and 10 m3, the consumer pays the same amount of R$ 9.00 (note that the subscription value must be paid even if the consumer has not consumed water). Above 10 m3, each subsequent cubic meter has a different value, depending on the consumption range. The table of values ​​per cubic meter for each consumption range is as follows:

up to 10 m3 - included in the franchise
11 to 30 m3 - R$ 1.00 per cubic meter
31 to 100 m3 -R$ 2.00 per cubic meter
101 m3 onwards - R$ 5.00 per cubic meter"