Create attendanceList
This code returns a count of students based in this problem:

"A teacher was unable to attend class on Monday and left a video for students to watch. To measure attendance, students were asked to place their registration number on an attendance list, and then enter the information into the system. Some students, however, as they have friends who run away from class, decided to be friends and entered the registration numbers of their runaway friends. The problem is that many students are friends with students who skip class and some enrollment numbers end up being repeatedly entered into the attendance list.

The teacher, upon noticing that the attendance list contained some repeated numbers, was at a loss, but decided to give a vote of confidence and give attendance to all the students whose enrollment numbers were on the list. As there are many students in the class and many numbers with repetition, she asked for your help to determine the total number of students who were present in the class."