Create divers
This program returns the identification of divers who did not return from diving, based in this problem: 

"An earthquake hit a city and affected many buildings, with the main containment dike having significant damage to its underground part and must be repaired quickly. The repair was carried out by divers, and each of them received a plate with an identification number at the beginning of the dive.

The dike was repaired, but unfortunately not all divers returned from the mission. You were hired to check who didn't return, after analyzing the license plates of those who did."